Ajout Team et Respawn
Ajout d'un système d'équipe, de respawn, de point de spawn (ca marche mais pas sûr à 100% y a toutes les fonctionnalité compliquées que je veux), J'ai aussi repassé le système de dégâts en int, et j'ai synchronisé les barres de vie et les morts !

+++ Assets/Scripts/Attaque/AutoAttaqueRanged.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class AutoAttaqueRanged : Photon.PunBehaviour {
 
-    public float impactDamage = 10;
-    public float splashDamage = 0;
+    public int impactDamage = 10;
+    public int splashDamage = 0;
     public float AOERadius = 0;
     public float fireRate = 1;
     public float projectileSpeed = 10;
